% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiveThirtyEightParser.R
\name{fiveThirtyEight_execute}
\alias{fiveThirtyEight_execute}
\title{Function to Scrape Five Thirty Eight Data and output as a CSV}
\usage{
fiveThirtyEight_execute()
}
\value{
csv file 'FiveThirtyEightParser.csv'
}
\description{
Function to Scrape Five Thirty Eight Data and output as a CSV
}
\examples{
fiveThirtyEight_execute()
}
